Unicode.Char.Identifiers

Description

Unicode Identifier and Pattern Syntax property functions based on Unicode Standard Annex #31

Documentation

isIDContinue :: Char -> Bool

Returns True if a character is an identifier continue character.

Since: 0.2.0

isIDStart :: Char -> Bool

Returns True if a character is an identifier start character.

Since: 0.2.0

isXIDContinue :: Char -> Bool

Returns True if a character is an identifier continue character, using the NFKC modifications detailed in UAX #31, 5.1.

Since: 0.2.0

isXIDStart :: Char -> Bool

Returns True if a character is an identifier start character, using the NFKC modifications detailed in UAX #31, 5.1.

Since: 0.2.0

isPatternSyntax :: Char -> Bool

Returns True if a character is a pattern syntax character.

Since: 0.2.0

isPatternWhitespace :: Char -> Bool

Returns True if a character is a pattern whitespace character.

Since: 0.2.0