Unicode.Char.Numeric.Compat

Description

Compatibility module for numeric character property related functions.

Since: 0.3.1

Predicates

isNumber :: Char -> Bool

Selects Unicode numeric characters, including digits from various scripts, Roman numerals, et cetera.

This function returns True if its argument has one of the following GeneralCategorys, or False otherwise:

DecimalNumber

LetterNumber

OtherNumber

Note: a character may have a numeric value (see numericValue) but return False, because isNumber only tests GeneralCategory: some CJK characters are OtherLetter and do have a numeric value. Use isNumeric to cover those cases as well.

isNumber c == Data.Char.isNumber c

@since 0.3.1 moved to Compat module.

Since: 0.3.0